requirements.txt
torch==2.6.0
torchvision==0.21.0
ultralytics==8.3.72
fastapi==0.115.8
uvicorn==0.34.0
pillow==11.1.0
python-dotenv==1.0.1
numpy==2.1.1
opencv-python-headless==4.11.0.86
python-multipart==0.0.20
pydantic==2.10.6
starlette==0.45.3
requests==2.32.3
tqdm==4.67.1

